Created Table Component and Wrote method Called Get Tables in TableRepository
