+++ Famicom/Components/Pages/Table.razor.cs
@@ -9,10 +9,6 @@ namespace Famicom.Components.Pages
 {
     public partial class TableBase : ComponentBase
     {
-        // Title and Image URL
-        public string TableTitle { get; set; } = "Manage Your Table";
-        public string TableImageUrl { get; set; } = ""; //image URL
-
         private TableModel tableModel {get; set; } = new TableModel();
         public ITable Table { get; set; }
 
@@ -20,17 +16,6 @@ namespace Famicom.Components.Pages
         {
             tableModel = new TableModel();
             Table = tableModel.GetTable();
-        }
-
-        // Methods for moving the table up and down
-        public void MoveTableUp()
-        {
-            Debug.WriteLine("Table moved up");
-        }
-
-        public void MoveTableDown()
-        {
-            Debug.WriteLine("Table moved down");
         }
     }
 }
